Monthly now supports cron expressions for habbit tracker
Also some improvements around Monthly, it wasn't naming files right
(went with human readable not integer value of month) also wasn't
putting the right name of the month at the top of the file.
Day letter also prints next to day value. Will be helpful for when we
get calendar stuff implemented as most the date work is done for that now